`[data-turbo-method="get"]` links: search params
Closes [hotwired/turbo#820][]
Follow-up to [hotwired/turbo#461][]

[hotwired/turbo#820]: https://github.com/hotwired/turbo/issues/820
[hotwired/turbo#461]: https://github.com/hotwired/turbo/pull/461

The background
---

According to the HTML Specification's [§ 4.10.21.3 Form submission
algorithm][] section, submissions transmitted as `GET` requests [mutate
the `[action]` URL][mutate], overriding any search parameters already
encoded into the `[action]` value:

> [Mutate action URL][algorithm]
> ---
>
> Let <var>pairs</var> be the result of converting to a list of
> name-value pairs with <var>entry list</var>.
>
> Let <var>query</var> be the result of running the
> `application/x-www-form-urlencoded` serializer with <var>pairs</var>
> and <var>encoding</var>.
>
> Set <Var>parsed action</var>'s query component to <var>query</var>.
>
> Plan to navigate to <var>parsed action</var>.

[§ 4.10.21.3 Form submission algorithm]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#form-submission-algorithm
[algorithm]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-mutate-action
[mutate]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#form-submission-algorithm:submit-mutate-action

Form submissions made with `POST` requests, on the other hand, encode
_both_ the `[action]` value's query parameters and any additionally
encoded body data:

> [Submit as entity body][post-submit]
> ---
>
> …
>
> Plan to navigate to a new request whose URL is <var>parsed
> action</var>, method is <var>method</var>, header list is
> « (`Content-Type`, mimeType) », and body is <var>body</var>.

[post-submit]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-body

The problem
---

When navigating through an `<a data-turbo-method="get">` with an
`[href]` that contains search parameters, the `<form>` element that
Turbo Drive constructs behind the scenes will unexpectedly omit the
search parameters during the transformation from `<a>` to `<form>`.

For example:

```html
<a href="/path-with-search-params?a=one&b=two"
   data-turbo-method="get">
  Submits as a form[method=get]
</a>
```

Would be translated to something like:

```html
<form method="get" action="/path-with-search-params?a=one&b=two" hidden></form>
```

Then, when submitting the `<form>`, the browser omits the `?one&b=two`.

The solution
---

This commit extends the current `FormLinkClickObserver` implementation
to loop over each entry in [URLSearchParams][] instance derived from the
provided [URL][] instance and transforms the name-value pair into an
[`<input type="hidden">`][input-hidden] element, which it appends to the
`<form>`.

Once that translation is complete, it removes any search parameters from
the form's `[action]` attribute to avoid duplicate entries in the form's
data.

While this change resolves the underlying issue, it's worth emphasizing
that declaring `[data-turbo-method="get"]` on an `<a>` element is
redundant, and should be avoided in the first place. Even though that
pattern should be avoided, this commit aims to prevent surprising
behavior whenever it occurs.

[URLSearchParams]: https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams
[URL]: https://developer.mozilla.org/en-US/docs/Web/API/URL
[input-hidden]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/hidden