Only snapshot frame visits that have actions
Frames that use `data-turbo-action` to promote their navigations to page
visits require cache snapshots to be taken prior to them loading, in
order for those snapshots to be accurate.

However snapshots can be expensive to create for complex pages. In cases
where a frame is *not* being promoted, we can avoid taking that
snapshot.

This helps with some performance issues observed when using frames on
very large pages.