Specify breakoutable frames in requesting page
Rather than use a meta tag in the response, we can put a central list of
"breakoutable" paths in a single meta tag, and include that on
requesting pages.

Assuming that an application will have a small number of these paths, it
might be simpler to centralise them in one place and include that in a
shared layout.