Allow changing the submitter text during form submission
This change introduces a new optional data-turbo-submitting text
attribute that can be set on submit elements (inputs or buttons) in
Turbo forms.

      <form action="/" method="post">
        <input type="submit" value="Save" data-turbo-submitting-text="Saving...">
      </form>

When the form is in a submmiting state Turbo will change the submitter
content -the input value for inputs, and the innerHTML for buttons- with
the data-turbo-submitting-text, and restore the original value when the
submission ends.

This is a common requirement in many apps, to style the form submitting
state and give feedback to the user that a click is already being
processed.