Form submissions from frames should clear cache
Usually when submitting a form with a method other than `GET`, we clear
the snapshot cache. This helps to avoid cases where we might show a
stale view of some data that we just mutated.

Form submissions from inside frames were not doing this. But they ought
to, for the same reason: that form submission may have mutated some data
that causes other recent pages to become stale.

To avoid this, we can clear the cache after processing a non-`GET` form,
or after any failed form submission, to mirror what we do when the form
is outside of any frames.