Rename `isIdempotent` to `isSafe`
The method name was a bit misleading, as what we're actually checking
for here is whether the method is [safe][0].

We can also now make use of this method in a couple of places where we
clear the snapshot cache after form submissions. The logic there is to
clear the cache when performing an operation that is unsafe.

[0]: https://developer.mozilla.org/en-US/docs/Glossary/Safe/HTTP