Remove `SubmitEvent` polyfill
The codebase's [SubmitEvent][] polyfill code involves three parts:

* a general purpose polyfill for browsers that do not provide a
  `window.SubmitEvent` prototype that includes a `submitter` property
  for their [submit][] events
* a Safari-specific bugfix
* a TypeScript `type` declaration.

As of April 2023, all facets of the `SubmitEvent` have fully support and
[Compatibility][] with evergreen browsers.

As of 2021-08-30, the [Safari bug][] mentioned in the code comment has
been resolved, and a fix has been released as part of Safari 15.4.

The TypeScript compiler supports the `SubmitEvent` type declaration.

This commit removes the polyfill from the codebase, and relies on the
built-in support instead.

[submit]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/submit_event
[SubmitEvent]: https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent
[Compatibility]: https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent#browser_compatibility
[Safari bug]: https://bugs.webkit.org/show_bug.cgi?id=229660