Fix frame not advancing when using `Turbo.visit`
When you have a frame with `data-trubo-action` attribute and use
`Turbo.visit({frame: "frame-id"})` to navigate the frame the action is
ignored and the url does not change.

This appears to happen as the `visit` method of the frame does not set
the `action` on the `FrameController` as it does in the
`proposeVisitIfNavigatedWithAction`.

To fix this I've set the `action` property on the frame in the visit
method.