Perform view transitions on cached snapshots
Extends the recently-added support for view transitions to better
support cached snapshot previews.

Whenever Turbo navigates to a page with a cached snapshot, we actually
render the page twice: first with the preview, then with the freshly
loaded version. To prevent this extra rendering from interfering with a
view transition, we previously disabled snapshot caching for pages that
had opted in to using a view transition. This keeps the transitions
uninterrupted, but has the disadvantage of preventing caching in cases
where we might still want it.

We can provide a better experience by allowing caching as normal, but
performing the view transition on whichever render happens first. In the
case of a cached snapshot, this allows us to fetch the real content
while we're transitioning to the cached view; when the new content is
ready and the transition has complete, we swap the new content into
place.

This also means that history navigations can still be served entirely
from cache, as they would be when not using view transitions.

In cases where a particular transition looks better without the
snapshot, we can use Turbo's existing mechanisms to opt out of caching
on those specific pages.