Rename variables to communicate their class
The `FormSubmission` delegate methods refer to arguments of type
`FetchRequest` as `request`, and arguments of `FetchResponse` as
`response`. Without type annotations to go along with those names, it
isn't clear that `request` is not a built-in `Request` and `response` is
not a built-in `response`.

This commit renames them to communicate their underlying types.