Add disk cache store (#949)
This commit extends the Turbo cache API to allow for the use of
different cache stores. The default store is still the in-memory store,
but a new persistent cache store is now available.

The disk store uses the [Cache API](https://developer.mozilla.org/en-US/docs/Web/API/CacheStorage)
to store and retrieve snapshots from the browser's cache. This allows
for the snapshots to be persisted across different tabs, page reloads
and even browser restarts.

The disk store is not enabled by default. To enable it, you need to set
the `Turbo.cache.store` property to `"disk"`.

```js
Turbo.cache.store = "disk"
```

This is also a stepping stone to implement offline support with Service
Workers. With a Service Worker in place, and the disk cache store
enabled, we can still serve cached snapshots even when the browser is
offline.