Add test coverage for `StreamActions` export
According to the [official documentation][], applications that need to
extend `StreamActions` should be importing it directly from the module
instead of accessing it through `window.Turbo.StreamActions`.

This commit adds test coverage to exercise the corresponding `import`.
It also re-arranges the `export` to explicitly occur in the
`@hotwired/turbo` module itself, instead of being transitively exported
as part of the `@hotwired/turbo/core` module.

[official documentation]: https://turbo.hotwired.dev/handbook/streams#custom-actions