Replace global `fetch` patch with Turbo-specific behavior
As an alternative to patching `fetch` globally, this commit introduces a
new `FetchRequestObserver` class, and makes the `Session` instance a
delegate.

As prep-work, this commit also moves the `recentRequests` out of the
global `@hotwired/turbo` module an into an instance property on
`Session`.

The `FetchRequestObserver` instance attaches an event listener for
`turbo:before-fetch-request` (in this case, on the `<html>` element).
Instead of generating the request ID in the delegate, this commit also
extends the `FetchRequest` class to generate the `uuid()` and set it to
the new `FetchRequest.id` property.

The `Session.willFetch` delegate method will be invoked before each
request, and will read the `X-Turbo-Request-Id` header and write it to
the instance's `recentRequests` set.