Set `html[lang]` during navigation
Closes [#967][]

Change the `<html>` element's `[lang]` attribute during navigation.
Currently, this isn't possible from applications (without a lot of
trouble), since the [turbo:before-render][] event is scoped to the
`newBody`.

In order to support this behavior, change the `PageSnapshot` constructor
to accept an instance of [HTMLHtmlElement][] (the `<html>` element) so
that that element can be accessible from the `PageRenderer`.

[#967]: https://github.com/hotwired/turbo/issues/967
[turbo:before-render]: https://turbo.hotwired.dev/reference/events
[HTMLHtmlElement]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLHtmlElement