Restore attribute "refresh=morph" to flag turbo frames to reload during a page refresh.
This adds a new turbo-frame attribute: `refresh`. When its value is `morph`:

* It will reload the turbo frame with morphing during a page refresh.
* It won't update the turbo frame with the server response.
* It won't remove it if it's missing in the server response.

This attribute was part of the original proposal we presented in Rails World, then
we removed it [1], because we thought it wasn't needed. But after testing
the library in different scenarios, we've found assuming certain behavior for all
the remote frames was problematic since it implied being too clever about what
you wanted to do with the frame. The new attribute makes for a simple behavior:

* The default behavior will be the expected one: turbo frames will be morphed
as any other element. If they get a new URL they will be reloadded, if the get
deleted in the response, they will disappear, etc.
* You can use the new attribute to flag the frames for which you want the special
behavior.

[1]: https://github.com/hotwired/turbo/pull/1019/commits/0c6a95d07fd7661e734ed1fbfe19daa9a00b3fd4